Phase 2 - wrote JsonReaderTest tests

+++ src/main/model/Goal.java
@@ -35,7 +35,7 @@ public class Goal implements Writable {
     }
 
     // EFFECTS: returns progress
-    public double getProgress() {
+    public int getProgress() {
         return progress;
     }
